requirements.txt
nats-py>=2.7.0
pydantic>=2.7.0
python-dotenv>=1.0.1
litellm>=1.40.0
mcp>=1.26.0
pytest>=8.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sse-starlette>=2.0.0
textual>=3.0.0
httpx>=0.27.0
prompt_toolkit>=3.0.0
rich>=13.0.0
psutil>=5.9.0
